Enabled tests around Fragments (from support library)
