+++ AndroidAnnotations/functional-test-1-5/AndroidManifest.xml
@@ -68,6 +68,7 @@
         <activity android:name="org.androidannotations.test15.nonconfiguration.NonConfigurationActivity_" />
         <activity android:name="org.androidannotations.test15.sherlock.MySherlockActivity_" />
         <activity android:name="org.androidannotations.test15.efragment.MyFragmentActivity_" />
+        <activity android:name="org.androidannotations.test15.efragment.MySupportFragmentActivity_" />
         <activity android:name="org.androidannotations.test15.afterviews.AfterViewsActivity_" />
         <activity android:name=".SSLConnection_" />
         <activity android:name="org.androidannotations.test15.SeekBarChangeListenedActivity_" />
@@ -82,7 +83,7 @@
 
         <service
             android:name="org.androidannotations.test15.eservice.MyService_"
-            exported="false" />
+            android:exported="false" />
     </application>
 
 </manifest>
