+++ AndroidAnnotations/functional-test-1-5/res/layout/support_fragments.xml
@@ -22,9 +22,9 @@
     android:orientation="horizontal" >
 
     <fragment
-        android:id="@+id/myFragment"
-        android:tag="myFragmentTag"
-        android:name="org.androidannotations.test15.efragment.MyFragment_"
+        android:id="@+id/mySupportFragment"
+        android:tag="mySupportFragmentTag"
+        android:name="org.androidannotations.test15.efragment.MySupportFragment_"
         android:layout_width="0dp"
         android:layout_height="fill_parent"
         android:layout_weight="1" />
